+++ martin/src/cog/mod.rs
@@ -1,5 +1,6 @@
 mod config;
 mod errors;
+mod image;
 mod model;
 mod source;
