+++ martin/src/cog/source.rs
@@ -1,18 +1,18 @@
 use std::collections::HashMap;
 use std::fmt::Debug;
 use std::fs::File;
-use std::io::BufWriter;
 use std::path::{Path, PathBuf};
 use std::vec;
 
 use async_trait::async_trait;
 use log::warn;
 use martin_tile_utils::{Format, TileCoord, TileInfo};
-use tiff::decoder::{ChunkType, Decoder, DecodingResult};
+use tiff::decoder::{ChunkType, Decoder};
 use tiff::tags::Tag::{self, GdalNodata};
 use tilejson::{TileJSON, tilejson};
 
 use super::CogError;
+use super::image::Image;
 use super::model::ModelInfo;
 use crate::file_config::{FileError, FileResult};
 use crate::{MartinResult, Source, TileData, UrlQuery};
@@ -27,8 +27,7 @@ struct Meta {
     origin: [f64; 3],
     // [minx, miny, maxx, maxy] in its model space coordinate system
     extent: [f64; 4],
-    zoom_and_ifd: HashMap<u8, usize>,
-    zoom_and_tile_across_down: HashMap<u8, (u32, u32)>,
+    images: HashMap<u8, Image>,
     nodata: Option<f64>,
 }
 
@@ -44,7 +43,15 @@ pub struct CogSource {
 impl CogSource {
     pub fn new(id: String, path: PathBuf) -> FileResult<Self> {
         let tileinfo = TileInfo::new(Format::Png, martin_tile_utils::Encoding::Uncompressed);
-        let meta = get_meta(&path)?;
+        let tif_file =
+            File::open(&path).map_err(|e: std::io::Error| FileError::IoError(e, path.clone()))?;
+        let mut decoder = Decoder::new(tif_file)
+            .map_err(|e| CogError::InvalidTiffFile(e, path.clone()))?
+            .with_limits(tiff::decoder::Limits::unlimited());
+        let model = ModelInfo::decode(&mut decoder, &path);
+        verify_requirements(&mut decoder, &model, &path.clone())?;
+
+        let meta = get_meta(model.clone(), &mut decoder, &path)?;
         let tilejson = tilejson! {
             tiles: vec![],
             minzoom: meta.min_zoom,
@@ -71,7 +78,7 @@ impl CogSource {
             Decoder::new(tif_file).map_err(|e| CogError::InvalidTiffFile(e, self.path.clone()))?;
         decoder = decoder.with_limits(tiff::decoder::Limits::unlimited());
 
-        let ifd = self.meta.zoom_and_ifd.get(&(xyz.z)).ok_or_else(|| {
+        let image = self.meta.images.get(&(xyz.z)).ok_or_else(|| {
             CogError::ZoomOutOfRange(
                 xyz.z,
                 self.path.clone(),
@@ -80,63 +87,8 @@ impl CogSource {
             )
         })?;
 
-        decoder
-            .seek_to_image(*ifd)
-            .map_err(|e| CogError::IfdSeekFailed(e, *ifd, self.path.clone()))?;
-
-        let (across, down) = self
-            .meta
-            .zoom_and_tile_across_down
-            .get(&(xyz.z))
-            .ok_or_else(|| {
-                CogError::ZoomOutOfRange(
-                    xyz.z,
-                    self.path.clone(),
-                    self.meta.min_zoom,
-                    self.meta.max_zoom,
-                )
-            })?;
-        let tile_idx;
-        if let Some(idx) = get_tile_idx(xyz, *across, *down) {
-            tile_idx = idx;
-        } else {
-            return Ok(Vec::new());
-        }
-        let decode_result = decoder
-            .read_chunk(tile_idx)
-            .map_err(|e| CogError::ReadChunkFailed(e, tile_idx, *ifd, self.path.clone()))?;
-        let color_type = decoder
-            .colortype()
-            .map_err(|e| CogError::InvalidTiffFile(e, self.path.clone()))?;
-
-        let (tile_width, tile_height) = decoder.chunk_dimensions();
-        let (data_width, data_height) = decoder.chunk_data_dimensions(tile_idx);
-
-        //do more research on the not u8 case, is this the right way to do it?
-        let png_file_bytes = match (decode_result, color_type) {
-            (DecodingResult::U8(vec), tiff::ColorType::RGB(_)) => rgb_to_png(
-                vec,
-                (tile_width, tile_height),
-                (data_width, data_height),
-                3,
-                self.meta.nodata.map(|v| v as u8),
-                &self.path,
-            ),
-            (DecodingResult::U8(vec), tiff::ColorType::RGBA(_)) => rgb_to_png(
-                vec,
-                (tile_width, tile_height),
-                (data_width, data_height),
-                4,
-                self.meta.nodata.map(|v| v as u8),
-                &self.path,
-            ),
-            (_, _) => Err(CogError::NotSupportedColorTypeAndBitDepth(
-                color_type,
-                self.path.clone(),
-            )),
-            // do others in next PRs, a lot of disscussion would be needed
-        }?;
-        Ok(png_file_bytes)
+        let bytes = image.get_tile(&mut decoder, xyz, &self.path)?;
+        Ok(bytes)
     }
 }
 
@@ -165,75 +117,6 @@ impl Source for CogSource {
     ) -> MartinResult<TileData> {
         Ok(self.get_tile(xyz)?)
     }
-}
-
-fn get_tile_idx(xyz: TileCoord, across: u32, down: u32) -> Option<u32> {
-    if xyz.y >= down || xyz.x >= across {
-        return None;
-    }
-
-    let tile_idx = xyz.y * across + xyz.x;
-    if tile_idx >= across * down {
-        return None;
-    }
-    Some(tile_idx)
-}
-
-fn rgb_to_png(
-    vec: Vec<u8>,
-    (tile_width, tile_height): (u32, u32),
-    (data_width, data_height): (u32, u32),
-    chunk_components_count: u32,
-    nodata: Option<u8>,
-    path: &Path,
-) -> Result<Vec<u8>, CogError> {
-    let is_padded = data_width != tile_width || data_height != tile_height;
-    let need_add_alpha = chunk_components_count != 4;
-
-    let pixels = if nodata.is_some() || need_add_alpha || is_padded {
-        let mut result_vec = vec![0; (tile_width * tile_height * 4) as usize];
-        for row in 0..data_height {
-            'outer: for col in 0..data_width {
-                let idx_chunk =
-                    row * data_width * chunk_components_count + col * chunk_components_count;
-                let idx_result = row * tile_width * 4 + col * 4;
-                for component_idx in 0..chunk_components_count {
-                    if nodata.eq(&Some(vec[(idx_chunk + component_idx) as usize])) {
-                        //This pixel is nodata, just make it transparent and skip it then
-                        let alpha_idx = (idx_result + 3) as usize;
-                        result_vec[alpha_idx] = 0;
-                        continue 'outer;
-                    }
-                    result_vec[(idx_result + component_idx) as usize] =
-                        vec[(idx_chunk + component_idx) as usize];
-                }
-                if need_add_alpha {
-                    let alpha_idx = (idx_result + 3) as usize;
-                    result_vec[alpha_idx] = 255;
-                }
-            }
-        }
-        result_vec
-    } else {
-        vec
-    };
-    let mut result_file_buffer = Vec::new();
-    {
-        let mut encoder = png::Encoder::new(
-            BufWriter::new(&mut result_file_buffer),
-            tile_width,
-            tile_height,
-        );
-        encoder.set_color(png::ColorType::Rgba);
-        encoder.set_depth(png::BitDepth::Eight);
-        let mut writer = encoder
-            .write_header()
-            .map_err(|e| CogError::WritePngHeaderFailed(path.to_path_buf(), e))?;
-        writer
-            .write_image_data(&pixels)
-            .map_err(|e| CogError::WriteToPngFailed(path.to_path_buf(), e))?;
-    }
-    Ok(result_file_buffer)
 }
 
 fn verify_requirements(
@@ -316,76 +199,82 @@ fn verify_requirements(
 }
 
 #[allow(clippy::cast_possible_truncation)]
-fn get_meta(path: &PathBuf) -> Result<Meta, FileError> {
-    let tif_file = File::open(path).map_err(|e| FileError::IoError(e, path.clone()))?;
-    let mut decoder = Decoder::new(tif_file)
-        .map_err(|e| CogError::InvalidTiffFile(e, path.clone()))?
-        .with_limits(tiff::decoder::Limits::unlimited());
-    let model = ModelInfo::decode(&mut decoder, path);
+fn get_meta(model: ModelInfo, decoder: &mut Decoder<File>, path: &Path) -> Result<Meta, FileError> {
+    let nodata: Option<f64> = if let Ok(no_data) = decoder.get_tag_ascii_string(GdalNodata) {
+        no_data.parse().ok()
+    } else {
+        None
+    };
     let origin = get_origin(
         model.tie_points.as_deref(),
         model.transformation.as_deref(),
         path,
     )?;
-    let (full_width_pixel, full_length_pixel) = decoder.dimensions().map_err(|e| {
-        CogError::TagsNotFound(
-            e,
-            vec![Tag::ImageWidth.to_u16(), Tag::ImageLength.to_u16()],
-            0, // we are at ifd 0, the first image, haven't seek to others
-            path.clone(),
-        )
-    })?;
-    let full_resolution = get_full_resolution(
+    let (full_width_pixel, full_length_pixel) = dim_in_pixel(decoder, path, 0)?;
+    let (full_width, full_length) = dim_in_model(
+        decoder,
+        path,
+        0,
         model.pixel_scale.as_deref(),
         model.transformation.as_deref(),
-        path,
     )?;
-    let full_width = full_resolution[0] * f64::from(full_width_pixel);
-    let full_length = full_resolution[1] * f64::from(full_length_pixel);
     let extent = get_extent(
         &origin,
         model.transformation.as_deref(),
         (full_width_pixel, full_length_pixel),
         (full_width, full_length),
     );
-    verify_requirements(&mut decoder, &model, path)?;
-    let mut zoom_and_ifd: HashMap<u8, usize> = HashMap::new();
-    let mut zoom_and_tile_across_down: HashMap<u8, (u32, u32)> = HashMap::new();
+    let mut images = vec![];
 
-    let nodata: Option<f64> = if let Ok(no_data) = decoder.get_tag_ascii_string(GdalNodata) {
-        no_data.parse().ok()
-    } else {
-        None
-    };
+    let mut ifd_idx = 0;
 
-    let images_ifd = get_images_ifd(&mut decoder, path);
+    loop {
+        let is_image = decoder
+            .get_tag_u32(Tag::NewSubfileType)
+            .map_or_else(|_| true, |v| v & 4 != 4); // see https://www.verypdf.com/document/tiff6/pg_0036.htm
+        if is_image {
+            //todo We should not ignore mask in the next PRs
+            let (tiles_across, tiles_down) = get_grid_dims(decoder, path, ifd_idx)?;
+            let image = Image {
+                ifd: ifd_idx,
+                across: tiles_across,
+                down: tiles_down,
+                nodata,
+            };
 
-    for (idx, image_ifd) in images_ifd.iter().enumerate() {
-        decoder
-            .seek_to_image(*image_ifd)
-            .map_err(|e| CogError::IfdSeekFailed(e, *image_ifd, path.clone()))?;
+            images.push(image);
+        } else {
+            warn!(
+                "A subfile of {} is ignored in the tiff file as Martin currently does not support mask subfile in tiff. The ifd number of this subfile is {}",
+                path.display(),
+                ifd_idx
+            );
+        }
 
-        let zoom = u8::try_from(images_ifd.len() - (idx + 1))
-            .map_err(|_| CogError::TooManyImages(path.clone()))?;
+        ifd_idx += 1;
 
-        let (tiles_across, tiles_down) = get_grid_dims(&mut decoder, path, *image_ifd)?;
-
-        zoom_and_ifd.insert(zoom, *image_ifd);
-        zoom_and_tile_across_down.insert(zoom, (tiles_across, tiles_down));
+        let next_res = decoder.seek_to_image(ifd_idx);
+        if next_res.is_err() {
+            //todo add warn!() here
+            break;
+        }
     }
-
-    if images_ifd.is_empty() {
-        Err(CogError::NoImagesFound(path.clone()))?;
-    }
-
+    let min_zoom = 0;
+    let max_zoom = (images.len() - 1) as u8;
+    let zoom_and_images: HashMap<u8, Image> = images
+        .iter()
+        .map(|image| {
+            let zoom = max_zoom.saturating_sub((image.ifd as u8) + 1);
+            (zoom, image.clone())
+        })
+        .collect();
     Ok(Meta {
-        min_zoom: 0,
-        max_zoom: images_ifd.len() as u8 - 1,
+        min_zoom,
+        max_zoom,
+        images: zoom_and_images,
         model,
         origin,
         extent,
-        zoom_and_ifd,
-        zoom_and_tile_across_down,
         nodata,
     })
 }
@@ -396,14 +285,14 @@ fn get_grid_dims(
     image_ifd: usize,
 ) -> Result<(u32, u32), FileError> {
     let (tile_width, tile_height) = (decoder.chunk_dimensions().0, decoder.chunk_dimensions().1);
-    let (image_width, image_length) = get_image_dims(decoder, path, image_ifd)?;
+    let (image_width, image_length) = dim_in_pixel(decoder, path, image_ifd)?;
     let tiles_across = image_width.div_ceil(tile_width);
     let tiles_down = image_length.div_ceil(tile_height);
 
     Ok((tiles_across, tiles_down))
 }
 
-fn get_image_dims(
+fn dim_in_pixel(
     decoder: &mut Decoder<File>,
     path: &Path,
     image_ifd: usize,
@@ -419,33 +308,22 @@ fn get_image_dims(
 
     Ok((image_width, image_length))
 }
+fn dim_in_model(
+    decoder: &mut Decoder<File>,
+    path: &Path,
+    image_ifd: usize,
+    pixel_scale: Option<&[f64]>,
+    transformation: Option<&[f64]>,
+) -> Result<(f64, f64), FileError> {
+    let (image_width_pixel, image_length_pixel) = dim_in_pixel(decoder, path, image_ifd)?;
 
-fn get_images_ifd(decoder: &mut Decoder<File>, path: &Path) -> Vec<usize> {
-    let mut res = vec![];
-    let mut ifd_idx = 0;
-    loop {
-        let is_image = decoder
-            .get_tag_u32(Tag::NewSubfileType)
-            .map_or_else(|_| true, |v| v & 4 != 4); // see https://www.verypdf.com/document/tiff6/pg_0036.htm
-        if is_image {
-            //todo We should not ignore mask in the next PRs
-            res.push(ifd_idx);
-        } else {
-            warn!(
-                "A subfile of {} is ignored in the tiff file as Martin currently does not support mask subfile in tiff. The ifd number of this subfile is {}",
-                path.display(),
-                ifd_idx
-            );
-        }
+    let full_resolution =
+        get_full_resolution(pixel_scale, transformation, path).map_err(FileError::from)?;
 
-        ifd_idx += 1;
+    let width_in_model = f64::from(image_width_pixel) * full_resolution[0].abs();
+    let length_in_model = f64::from(image_length_pixel) * full_resolution[1].abs();
 
-        let next_res = decoder.seek_to_image(ifd_idx);
-        if next_res.is_err() {
-            break;
-        }
-    }
-    res
+    Ok((width_in_model, length_in_model))
 }
 
 fn get_origin(
@@ -567,83 +445,10 @@ mod tests {
     use std::path::PathBuf;
 
     use insta::assert_yaml_snapshot;
-    use martin_tile_utils::TileCoord;
     use rstest::rstest;
     use tiff::decoder::Decoder;
 
     use crate::cog::model::ModelInfo;
-    use crate::cog::source::get_tile_idx;
-
-    #[test]
-    fn can_calc_tile_idx() {
-        assert_eq!(Some(0), get_tile_idx(TileCoord { z: 0, x: 0, y: 0 }, 3, 3));
-        assert_eq!(Some(8), get_tile_idx(TileCoord { z: 0, x: 2, y: 2 }, 3, 3));
-        assert_eq!(None, get_tile_idx(TileCoord { z: 0, x: 3, y: 0 }, 3, 3));
-        assert_eq!(None, get_tile_idx(TileCoord { z: 0, x: 1, y: 9 }, 3, 3));
-    }
-
-    #[rstest]
-    // the right half should be transprent
-    #[case(
-        "../tests/fixtures/cog/expected/right_padded.png",
-        (0,0,0,None),None,(128,256),(256,256)
-    )]
-    // the down half should be transprent
-    #[case(
-        "../tests/fixtures/cog/expected/down_padded.png",
-        (0,0,0,None),None,(256,128),(256,256)
-    )]
-    // the up half should be half transprent and down half should be transprent
-    #[case(
-        "../tests/fixtures/cog/expected/down_padded_with_alpha.png",
-        (0,0,0,Some(128)),None,(256,128),(256,256)
-    )]
-    // the left half should be half transprent and the right half should be transprent
-    #[case(
-        "../tests/fixtures/cog/expected/right_padded_with_alpha.png",
-        (0,0,0,Some(128)),None,(128,256),(256,256)
-    )]
-    // should be all half transprent
-    #[case(
-        "../tests/fixtures/cog/expected/not_padded.png",
-        (0,0,0,Some(128)),None,(256,256),(256,256)
-    )]
-    // all padded and with a no_data whose value is 128, and all the component is 128
-    // so that should be all transprent
-    #[case(
-        "../tests/fixtures/cog/expected/all_transprent.png",
-        (128,128,128,Some(128)),Some(128),(128,128),(256,256)
-    )]
-    fn test_padded_cases(
-        #[case] expected_file_path: &str,
-        #[case] components: (u8, u8, u8, Option<u8>),
-        #[case] no_value: Option<u8>,
-        #[case] (data_width, data_height): (u32, u32),
-        #[case] (tile_width, tile_height): (u32, u32),
-    ) {
-        let mut pixels = Vec::new();
-        for _ in 0..(data_width * data_height) {
-            pixels.push(components.0);
-            pixels.push(components.1);
-            pixels.push(components.2);
-            if let Some(alpha) = components.3 {
-                pixels.push(alpha);
-            }
-        }
-        let componse_count = if components.3.is_some() { 4 } else { 3 };
-        let png_bytes = super::rgb_to_png(
-            pixels,
-            (tile_width, tile_height),
-            (data_width, data_height),
-            componse_count,
-            no_value,
-            &PathBuf::from("not_exist.tif"),
-        )
-        .unwrap();
-        let expected = std::fs::read(expected_file_path).unwrap();
-        assert_eq!(png_bytes, expected);
-    }
-
     #[test]
     fn can_get_model_infos() {
         let path = PathBuf::from("../tests/fixtures/cog/rgb_u8.tif");
